[!] Freetown. Sync with [me_MemorySegment]..
For this case there is no some evil behemoth class and free functions
that are exploited by it.

We have two moderately disgusting export functions. Which occurred to
use subfunctions in their implementation.

According to zeitgeist vibe, those sub-functions are shown
in "Freetown" sub-namespace and thus are available for hire
by other modules.